# http/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Search Service API
  version: "1.0.0"
  description: >
    Provides full-text / fuzzy search over indexed JSON items
    using PostgreSQL pg_trgm strict word similarity scoring.
  contact:
    name: Search Service
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/search:
    get:
      operationId: searchItems
      summary: Search indexed items
      tags:
        - search
      description: >
        Returns a ranked list of items whose searchable values match the query
        string via pg_trgm strict word similarity.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string.
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (default 20, max 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination (default 0).
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Ranked search results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Bad request – missing or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/search/{asset_type}:
    get:
      operationId: searchItemsByType
      summary: Search items scoped to a single item type
      tags:
        - search
      description: >
        Returns a ranked list of items of the given type whose searchable
        values match the query string via pg_trgm strict word similarity.
      parameters:
        - name: asset_type
          in: path
          required: true
          description: The type of item to search within.
          schema:
            $ref: "#/components/schemas/ItemType"
        - name: q
          in: query
          required: true
          description: Search query string.
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (default 20, max 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination (default 0).
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Ranked search results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Bad request – missing or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/{asset_type}:
    get:
      operationId: listItems
      summary: List and filter items of a specific type
      tags:
        - search
      description: >
        Returns a paginated list of items for the given item type.
        Supports optional filtering by name, creator and license.
      parameters:
        - name: asset_type
          in: path
          required: true
          description: The type of item (lowercase).
          schema:
            $ref: "#/components/schemas/ItemType"
        - name: name
          in: query
          required: false
          description: Filter by item name (case-insensitive substring match).
          schema:
            type: string
        - name: creator
          in: query
          required: false
          description: Filter by creator name (case-insensitive substring match).
          schema:
            type: string
        - name: license
          in: query
          required: false
          description: Filter by exact license value.
          schema:
            $ref: "#/components/schemas/License"
        - name: sort
          in: query
          required: false
          description: Field to sort by (default "name").
          schema:
            $ref: "#/components/schemas/SortField"
        - name: order
          in: query
          required: false
          description: Sort direction (default "asc").
          schema:
            $ref: "#/components/schemas/SortOrder"
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (default 20, max 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination (default 0).
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItemsResponse"
        "400":
          description: Bad request – invalid parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/{asset_type}/{item_id}/thumbnail:
    get:
      operationId: downloadThumbnail
      summary: Download the thumbnail for an item
      tags:
        - storage
      description: >
        Streams the thumbnail image for the given item. Returns 404 if the
        item does not exist or has no thumbnail. Thumbnails are immutable,
        so the response includes Cache-Control and ETag headers. Clients
        can send If-None-Match to receive 304 Not Modified.
      parameters:
        - name: asset_type
          in: path
          required: true
          description: The type of item (lowercase).
          schema:
            $ref: "#/components/schemas/ItemType"
        - name: item_id
          in: path
          required: true
          description: Unique identifier of the item.
          schema:
            type: string
            format: uuid
        - name: If-None-Match
          in: header
          required: false
          description: >
            ETag value from a previous response. If the thumbnail has not
            changed, the server responds with 304 Not Modified.
          schema:
            type: string
      responses:
        "200":
          description: Thumbnail image content.
          headers:
            Cache-Control:
              schema:
                type: string
                example: "public, max-age=31536000, immutable"
            ETag:
              schema:
                type: string
                example: "\"019cf922-a443-79db-a67e-342b6edea3cf\""
          content:
            image/png:
              schema:
                type: string
                format: binary
        "304":
          description: >
            Not Modified – the client's cached version is still current.
        "404":
          description: Item or thumbnail not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ItemType:
      type: string
      enum:
        - map
        - gameskin
        - hud
        - skin
        - entity
        - theme
        - template
        - emoticon
    License:
      type: string
      description: |
        Content license. Supported values:
        - cc0          – Creative Commons Zero (public domain dedication)
        - cc-by        – CC Attribution
        - cc-by-sa     – CC Attribution-ShareAlike
        - cc-by-nd     – CC Attribution-NoDerivs
        - cc-by-nc     – CC Attribution-NonCommercial
        - cc-by-nc-sa  – CC Attribution-NonCommercial-ShareAlike
        - cc-by-nc-nd  – CC Attribution-NonCommercial-NoDerivs
        - gpl-2        – GNU General Public License v2
        - gpl-3        – GNU General Public License v3
        - mit          – MIT License
        - custom       – Custom / other known license
        - unknown      – License not specified
      enum:
        - cc0
        - cc-by
        - cc-by-sa
        - cc-by-nd
        - cc-by-nc
        - cc-by-nc-sa
        - cc-by-nc-nd
        - gpl-2
        - gpl-3
        - mit
        - custom
        - unknown
...
    SearchResult:
      type: object
      required: [item_id, asset_type, score, item_value]
      properties:
        item_id:
          type: string
          format: uuid
          description: Unique identifier of the item.
        asset_type:
          $ref: "#/components/schemas/ItemType"
        score:
          type: number
          format: double
          description: Relevance score (higher is better).
        item_value:
          type: object
          description: Arbitrary JSON payload of the item.
          additionalProperties: true
    SearchResponse:
      type: object
      required: [results, total]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
          description: Total number of results before pagination.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    ListItemsResponse:
      type: object
      required: [results, total]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ListItem"
        total:
          type: integer
          description: Total number of results before pagination.
    ListItem:
      type: object
      required: [item_id, asset_type, item_value]
      properties:
        item_id:
          type: string
          format: uuid
          description: Unique identifier of the item.
        asset_type:
          $ref: "#/components/schemas/ItemType"
        item_value:
          type: object
          description: Arbitrary JSON payload of the item.
          additionalProperties: true
    SortField:
      type: string
      description: Field to sort results by.
      enum:
        - name
        - created_at
      default: name
    SortOrder:
      type: string
      description: Sort direction.
      enum:
        - asc
        - desc
      default: asc